pyproject.toml
[project]
name = "furangel-invoice-parser"
version = "0.1.0"
description = "Furangel invoice parser"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "flask==3.1.0",
    "fuzzywuzzy==0.18.0",
    "google-api-core==2.24.1",
    "google-api-python-client==2.161.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "google-cloud-secret-manager==2.23.0",
    "google-genai>=1.13.0",
    "googleapis-common-protos==1.67.0",
    "grpc-google-iam-v1==0.14.0",
    "grpcio==1.70.0",
    "grpcio-status==1.70.0",
    "gunicorn==23.0.0",
    "httplib2==0.22.0",
    "idna==3.10",
    "iniconfig==2.0.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.5",
    "levenshtein==0.26.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "numpy==2.2.3",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pandas==2.2.3",
    "pluggy==1.5.0",
    "proto-plus==1.26.0",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pygments==2.19.1",
    "pyparsing==3.2.1",
    "pypdf==5.3.0",
    "pytest==8.3.4",
    "pytest-mock>=3.14.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-levenshtein==0.26.1",
    "pytz==2025.1",
    "rapidfuzz==3.12.1",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rsa==4.9",
    "six==1.17.0",
    "tzdata==2025.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "werkzeug==3.1.3",
]

[tool.pytest.ini_options]
pythonpath = "src"
